+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # Changelog
 
 ## Unreleased
+- Removed the DevSkill parsers because DevSkill is no longer an online judge
 - Fixed the parsing of CodeChef problems on its new problem page
 
 ## [2.33.0](https://github.com/jmerle/competitive-companion/releases/tag/2.33.0) (2022-04-03)
 - Added a parser for Newton School
-- Removed the DevSkill parsers because DevSkill is no longer an online judge
 - Fixed the parsing of CodeChef sample cases for problems where the input/output headers are not prefixed with "Example" or "Sample"
 - Fixed the parsing of CSES problems on CSES's www subdomain
 - Fixed the parsing of BUCTOJ problems and contests