+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Removed the DevSkill parsers because DevSkill is no longer an online judge
 - Fixed the parsing of CodeChef problems on its new problem page
+- Fixed the parsing of Virtual Judge sample cases where the sample cases are displayed in a code block that is nested in another code block
 
 ## [2.33.0](https://github.com/jmerle/competitive-companion/releases/tag/2.33.0) (2022-04-03)
 - Added a parser for Newton School